Add files via upload
Turing Machine Simulation for String Validation in {0,1} Language.Provides an in-depth understanding of how Turing machines work, including transition tables, states, and heads. Users can input a binary string consisting of only '0' and '1' characters, which will then be processed by the Turing machine based on the rules in the transition table. The Turing machine will perform a series of transitions between states to determine whether the string is valid or not.